+++ packages/block-library/src/gallery/edit.js
@@ -585,7 +585,7 @@ function GalleryEdit( props ) {
 					{ Platform.isNative ? (
 						<SelectControl
 							__nextHasNoMarginBottom
-							label={ __( 'Link to' ) }
+							label={ __( 'Link' ) }
 							value={ linkTo }
 							onChange={ setLinkTo }
 							options={ linkOptions }
@@ -630,7 +630,7 @@ function GalleryEdit( props ) {
 				<BlockControls group="block">
 					<ToolbarDropdownMenu
 						icon={ linkIcon }
-						label={ __( 'Link To' ) }
+						label={ __( 'Link' ) }
 					>
 						{ ( { onClose } ) => (
 							<MenuGroup className="blocks-gallery__link-to-control">